Move key fetching code from main.go to a separate file
